大版本修改

+++ zefu-modules/zefu-business/src/main/java/com/zefu/business/cache/IProductFuncCache.java
@@ -70,5 +70,5 @@ public interface IProductFuncCache {
      * @param funcStatus 发布状态 null则查询所有
      * @return
      * */
-    List<BusProductFunc> getListProperties(String productCode, String funcType, Long funcStatus);
+    List<BusProductFunc> getListProperties(String productCode, String funcType, Integer funcStatus);
 }
